requirements.txt
pytest>=8.0.0
pytest-asyncio>=0.23.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
tenacity>=8.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4